Move all products to dedicated folder and use more pomless and clean up
Remove the org.eclipse.equinox.p2.reconciler.dropins/Bootstrap.product,
which is not used anywhere and seems only to be used to boot-strap the
'Reconciler' launch-configs. They are outdated and can be simpler
created from the launch-config dialog.

Remove now obsolete pom-files. They were only necessary because
Tycho-pomless otherwise does not yet use packaging-type
'eclipse-plugin'.

Remove p2.inf files for products. They add repositories that don't
exist. And if a update-repo is desired later, it can be specified
in the product-file.

Remove launch-configs associated to products. If one wants to launch
the product, it can be easily done from the Product-editor.