Move all products to dedicated folder and build them
Remove the org.eclipse.equinox.p2.reconciler.dropins/Bootstrap.product,
which is not used anywhere and seems only to be used to boot-strap the
'Reconciler' launch-configs. They are outdated and can be simpler
created from the launch-config dialog.

And remove now obsolete pom-files. They were only necessary because
Tycho-pomless otherwise does not use packaging-type 'eclipse-plugin'.

Remove p2.inf files for products, because they add repositories that
don't exist. And if a update-repo is desired later it can be configured
in the product.

Remove launch-configs associated to products. If one wants to launch
the product, it can be easily done from the Product-editor.