Check for PGP signatures even if the artifact is jar signed
For every artifact there must be at least one trusted certificate or one
trusted PGP key.  So if there is a trusted certificate there is no need
to check for PGP keys.  Also, if there is a trusted PGP key, then
regardless of whether there are untrusted certificates, the artifact is
trusted based on the PGP key's trust.

Use the information returned from the trust prompt to properly compute
whether there are any remaining untrusted artifacts.  In particular,
don't assume the UI service has properly done that computation.

Also verify that a jar-signed artifact's signature is within the
validity period of the certificate.

https://github.com/eclipse-equinox/p2/issues/121