Improve the handling of revoked PGP keys
In general, handle a revoked key as if all signatures that use it are
potentially tampered and require explicit review and additional extra
confirmation.

Check revocation during download in PGPSignatureVerifier, but only log
error rather than fail the download so that p2 does not try to
repeatedly download the same artifact because it assumes that it's
corrupted in some way. Revocation is handled later in the
CertificateChecker for presentation to the user.

DefaultPGPPublicKeyService.getVerifiedRevocationDate handles exceptions
when checking revocation as if the key were revoked:

https://bugs.eclipse.org/bugs/show_bug.cgi?id=581453

The CertificateChecker removes from the set of trusted keys all PGP keys
that are revoked.

https://bugs.eclipse.org/bugs/show_bug.cgi?id=581452

I.e., a revoked key can never be trusted, even if present in the trusted
preferences, but rather will always be presented to the user via the
TrustCertificateDialog.  The TrustCertificateDialog is made aware of
revoked PGP keys and presents detailed information to the users.
Moreover, accepting to trust revoked keys involves an extra confirmation
so that the user is explicitly aware of the risk.

The TrustPreferencePage also shows revocation information.