Provide support to eject untrusted content from the cache/pool
Enhance SimpleArtifactRepository so that it records the added
descriptors and can be asked to remove descriptors if they have been
added.

This allows CertificateChecker, in the case that the user rejects
untrusted content, to easily eject rejected artifacts from the pool.
This approach also ensures that for cases like a shared global bundle
pool,  artifacts are removed only if they were downloaded during the
current provisioning process, but not if they were already present
previously, in which case removing them would break existing
installations.

https://bugs.eclipse.org/bugs/show_bug.cgi?id=581451