Support tracking and reviewing content origin during installation
Tracking tracking is provided by ProvisioningContext's
getInstallableUnitSources and getArtifactSources which do a deep
analysis of all repositories, including children of composites and
repository reference to track installable units and artifact keys back
to their originating repository.

This is used during the Engine's Collect phase as implemented by the new
AuthorityChecker, which also manages underlying preferences for keeping
track of trusted authorities.

IInstallableUnitUIServices provide support for prompting the user and is
implemented by ValidationDialogServiceUI to display TrustAuthorityDialog
which is very similar in design to TrustCertificateDialog.

The TrustPreferencePage is enhanced to provide user-level access to the
trusted authorities preferences.

https://github.com/eclipse-equinox/p2/issues/235