CacheManager: Fix exception handling (#259)
Method

    createCacheFromFile()

was introduced in

    Bug 464614 - Use XZ as compression formats of metadata files

and is more a less a simplified copy of

    createCache()

which only deals with a single URI, instead of two (with and without
.jar appended).

However, the condensed exception handling was missing important branches
and swallowed too many exceptions. This caused that a
FileNotFoundException Was silently swallowed and the low-level Transport
was asked anyways, instead of bailing out with a proper
PrivisionException immediately.

The low-level Transport then, after sending another redundant HTTP
request, returned a ProvisionException with improper status code
(ARTIFACT_NOT_FOUND instead of REPOSITORY_NOT_FOUND).

Fixes #257.