Remove static Activator reference access
Currently two classes only access the "instance" of the Activator class
in a static way what often is a source of failures when using P2 in
other context as a potential NPE can occur.

This replaces the static access with some OSGi low level access that
will fullfill a similar purpose but do not require to tie these to the
activator class.