Discover ECF provided schemes
Currently if one tries to use a custom ecf scheme the first time the
install dialog complains that the scheme is not supported. A user has
first use some standard scheme (e.g. http) then ECF is loaded and the
custom scheme can be used.

This enhance the repository check in a way that it discovers ECF
available schemes at the first place.