Add IQueryable.contains(T) method and implement overrides where suitable
This methods provides a more convenient and in most cases faster way to
test if a queryable (e.g. a repository) contains an element, compared to
querying the queryable/repository for it.
IArtifactRepository already has a contains() method so this effectively
only adds it for IMetadataRepository and generalizes it for all
IQueryable sub-types.

Co-authored-by: Christoph Läubrich <laeubi@laeubi-soft.de>