Add options to the UI to disable and remove update sites for cases where update sites are not found (#207)
Previously, if accumulation of load failures was enabled, all failures were put into a MultiStatus and displayed to the user. This has been changed to use more specific dialogs if possible. If all load failures were caused by repositories not being found, dialogs that allow removing or disabling the affected sites are used. If there is only a single load failure and it was caused by a repository not being found, the "Error Contacting Site" dialog is used.

This commit extends the "Error Contacting Site" dialog with buttons for editing, removing or disabling a single site.