Add IInstallableUnitQueryable interface to allow easier adaption
Currently there is only the generic IQueryable<T> but this don't work
well in cases where one can't pass type information for the generic e.g.
when adapting an object there is no way to pass the parameter T.

This adds a new interface that lock in the type T for a IQueryable that
can supply IInstallableUnits.