Prevent trust dialogs from being hidden behind the application window
Use SWT.ON_TOP for the shell in order to keep the dialogs visible,
because generally the user cannot proceed with their activity
(update/install) until the dialog is completed.

A modal dialog is avoided because these dialogs are asynchronously
opened from a background job, which blocks, so there is no knowing what
UI activity might currently be under way.

https://github.com/eclipse-oomph/oomph/issues/11