Do content sensing to make sure the file is really not a jar
Currently we rely on the file-ending to determine the type of file but
this can fail in cases a CacheManger returns a cached file without that
extension (for whatever reason).

This now adds another check for content sensing if the simple check
fails to ensure a zipped item is not parsed as plain text.