Improve the formatting of DirectorApplication
When modifying DirectorApplication, the save actions generally transform
and reformat the source code because those actions where not applied in
the past. This makes it difficult to see what is being semantically
modified versus which is modified by sematnic-perserving the save
actions. Apply these changes now, once and for all.

https://github.com/eclipse-equinox/p2/issues/365