Improve RepositoryHelper.getSharedBundlePools to load all agent pools
Oomph has moved the encoding.info to be in the same folder as the
agents.info which lists all the agent locations (rather than in the
folder of the default agent's location). The agents.info file can be
loaded to determine the location of the pools.info of each agent. This
refined approach covers the full generality of Oomph's shared agent
manager allows all pools of all agents to be located with no assumptions
about defaults.

https://github.com/eclipse-equinox/p2/issues/373