Allow subclasses to control agent creation for DirectorApplication
Currently the DirectorApplication internally creates or fetch an agent
using the OSGi registry. If the application is used standalone (e.g.
Tycho) it is sometimes desirable to have more control over this process,
for this purpose there are now two new method to either fetch the
default agent or acquire a new one for a given location.