Make mirror parse errors configurable
Currently Tycho tests are quite unstable because it could happen that
mirrors list can't be processed and then an error occur in the logs what
make the test fail their "no errors".

But also as a user one might be annoyed to see errors because mirror
list can't be processed.

This adds a new static MIRROR_PARSE_ERROR_LEVEL filed to allow
configuration of the error level.