Save state should be reset to the previous state
Currently if one performs an operation as batch and then inside that
also performs a batch this resets the save flag even though there are
more items to batch in the operation.

This saves the current state and restore that state on completion, save
is only performed if saving was not already disabled before.