Refresh a subset of bundles if they provide missing requirements
When refreshing large sets of bundles the resolver sometimes take a
choice where a requirement can't be bound even though it is there and
resolvable.

This adds some additional code to detect this situation and then refresh
a smaller set of bundles that provide these until a max retry is
reached, all bundles are resolved, or all providers have been tried to
refresh already.