Add a filter to  requirements corresponding to a feature's imports
Provide new FeatureEntry.createRequires methods that specify isImport,
deprecating the older ones, and provide a field as well as accessors for
isImport. Use that attribute in FeaturesAction.getFilter(FeatureEntry)
to add a filter (!(org.eclipse.equinox.p2.exclude.import=true)) to the
synthesized requirement.

Update StringBuffer to StringBuilder in the otherwise-modified files.

https://github.com/eclipse-equinox/p2/issues/138