Only select IUs up to the max of a requirement using highest version
Currently if the slicer is given a set of IUs where multiple ones match
a requirement, all of them are selected independently of the actual
number of max constraint.

This is now changed by collecting all IUs that satisfy the requirement,
then sort them by highest version and restrict the result to the maximum
number per id.