Add optional, greedy and disabled by default requirement to the source
Currently it is quite cumbersome effort to include the source of a
bundle with an installation even though there might be multiple demands,
for example legal reasons or debugging.

This now adds a new greedy, optional requirement that is disabled by a
filter so it is not effective by default but enables an install agent to
specify a property of org.eclipse.update.install.source=true to
include sources of bundles in products whenever available.