Adjust maven metadata to the ones actually published to central
Currently the project is configured to use the org.eclipse.equinox.p2
group id but it is actually published to org.eclipse.platform. That is
both confusing and makes the metadata in p2 pointless and not allows to
use local build artifacts (e.g. for Tycho).

This changes the group id and configures the update-consumer-pom to
create a pom that can be used in a local maven setup similar to one
published on maven central.