Embed the required StateFactory implementations
Currently P2 has the BundleDescription in its API where the only
implementation is in the discouraged osgi.compatibility fragment, but
actually P2 does not really need a full implementation, it is just using
that to parse manifest headers and generate metadata from it.

This embeds a stripped down version of the actual implementation in the
p2.publisher.eclipse that allows to create implementation objects
without rely on the compatibility fragment anymore.