Use Map.(copy)of(...) in ProvidedCapability
in the context of Tycho it could easily happen that there are hundred of
thousands of ProvidedCapability's. Each of them currently carry a
HashMap that is then wrapped in an unmodifiable collection.

This replaces it with Map.copyOf(...) now to ensure it is unmodifiable
and only ever contains the exact number of entries saving a few bytes
for object references and possible null table entries.