Nullpointer in AbstractRepositoryManager#handleRemoteIndexFile
I found that in some rare cases the handleRemoteIndexFile could run into
a NPE if the Transport service can't be fetched.

This now checks first for null before using the transport and if it is
null fall back to simple URL stream, if that also fails, create an empty
index file.