Absence of an unpack attribute must not result in BundleShapeAdvice
Currently the FeaturesAction assumes that the unpack attribute is always
set, if that is not the case it leads to unexpected results because then
unpack=true is assumed always. This is especially dangerous as PDE since
2023-12 release removes the attributes and Tycho assume they are false
by default the same as PDE has done in the past when adding new items.

To mitigate this this adds an additional check to only evaluate the
unpack if it is set, together with a testcase that covers all possible
combinations.