Improve message if artifact can not be downloaded from any repository
There are rare cases where one gets a message that an artifact can not
be downloaded from any repository, this has two pitfalls:

1) You don't get an idea what repositories are actually contacted
2) Its hard to find out who actually requires this artifact

This now enhances an IArtifactRequest by a method to set a "context iu"
why this artifact is requested. If the information is given, it searches
through the available IUs and gives additional information why it is
required and what repositories are queried.