VersionRange for root features is too tight when expanded version given
Currently the (disabled by default) requirement for a root feature uses
an exact version range when the feature version is expanded (e.g. Tycho)
but the purpose of a root feature is that it can be updated what such a
strict range makes it impossible (if activated).

This now adds a new IVersionRangeAdvice that allows to modify the result
of a product included requirements version range.