Use RepositoryTracker for KnownRepositories to be consistent with Update
Currently if the user choose 'Check for Updates' the handler uses the
ProvisioningUI (that uses RepositoryTracker internally) to choose the
initial items to offer. The UpdateChecker (aka automatically search for
updates) on the other hand uses a (potentially) different set of items.

This now adds a new method that allow to pass the repository flags to be
used when search for an update and delegate it on the callers side to
use the Provisioning UI.