Support empty lower and upper bounds in P2 VersionRange
This enables support for version range specifications where the lower or
upper bound is not specified and the delimiting bracket is directly next
to the separating colon. For example '[1,)' or '(,1]' but also '(,)' can
then be parsed successfully.

The type of supported brackets for an empty or unspecified bound is
something that's not inherently obvious. From a mathematical perspective
an empty bound can be considered equivalent to infinity. And in math may
not say inclusive infinity. From that perspective only '[1,infinity)'
respectively '[1,)' would be permitted, but not '[1,]'.
But technically infinity has a specific value for P2 versions, i.e.
Version.MAX_VERSION. So if one really wants to cover all possible
values, the upper bound still has to be inclusive, i.e. ']'. For the
lower bound the situation is similar, where '0.0.0' or the
'empty'-version is the smallest one possible.