Migration to more consistent property name for provided service
Currently we use the property 'p2.agent.servicename' to identify the
provided service, but the constant is named PROP_CREATED_SERVICE_NAME.
So beside the inconsistency of created <> agent it shows that the "name"
part should actually be an own segment (like in AgentServiceFactory).
This also affects the naming of component property types

Because of this we introduce here a new constant PROP_AGENT_SERVICE_NAME
with the value 'p2.agent.service.name' and deprecate the old constant
while remain support for both properties.