Do not use a job in the FileInfoReader
The FileInfoReader currently creates a job just to join its execution,
as the job is also marked as a system job to hide it from the users this
does not seem very useful and only complicates the flow and increased
the workload.

This now removes the job extension and also making the monitor field
obsolete now.