Record Package Attributes and Directives on exported packages
Currently P2 lacks support for package attributes and directives that
can lead to wrong selection of providers. One issue is that the
information on exported packages are not recorded at all in the P2
metadata and therefore can not be used in slicer or planner operations.

As a first step this now records the information on the provider side by
transforming attributes and directives into properties of the package
capability.