Add BundlesAction.createBundleIU(File) for easy creation of IU
It is often useful to get some information about a file and
BundlesAction is the P2 way to gather such data, but at the moment it is
really cumbersome and requires a lot of calls and handling of cases to
get to that point.

This now adds a new method BundlesAction.createBundleIU(File) that ease
this task.