Use SWT.ON_TOP flag for MessageDialog created in TrustCertificateDialog
When clicking on the "Always trust all authorities" checkbox, with
Eclipse running on GTK+, the created MessageDialog is hidden by the
"main" dialog of TrustCertificateDialog. After this, the "main" dialog
cannot be interacted with using the mouse pointer, giving the impression
that Eclipse is frozen and unusable.

This change adds the SWT.ON_TOP flag to MessageDialog dialogs created in
TrustCertificateDialog, so that the question dialog appears above the
TrustCertificateDialog dialog - that way the question dialog is seen by
the user and can be interacted with normally.

Fixes: #570